pyproject.toml
[project]
name = "postsbackend"
version = "0.1.0"
description = "Backend for Posts app"
authors = [
    {name = "Balashov Igor",email = "igor2233balashov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "aiobotocore (>=2.22.0,<3.0.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "alembic (>=1.15.2,<2.0.0)",
    "redis (>=6.1.0,<7.0.0)",
    "fastapi-cache2-fork (>=2.2.0,<3.0.0) ; python_version >= \"3.12\" and python_version < \"4.0\"",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"

